Moved dendrogram functionality into self-contained plugin
